Wow 2.0.
I finished the SQLite implementation.

What does this mean? It means I can now create Notebook objects with their values, store them in an SQLite database (as in an actual file), and
retrieve them whenever needed. Not only that, but I also adjusted the NotebookSelection activity to read all the notebooks found in the SQL
database and then create a new NotebookSelectionCard for each of them. Right now, the app detects if it is set to debug mode (true by default),
checks the database to see if it already has a set of debug notebooks assigned, and if not the app will save a list onto the database consisting of
differnt simple notebooks I made. Since it was a pain to do each notebook individually, I wrote a python script that will randomly generate notebooks
for me on demand, and give me a line of code, ready to be pasted in.